* encrypt module importing moved to new api module, as there is problem while installing this product in virtualenv because of cryptography and asn1crypto dependencies. * python2/3 compatiblity added in encrypt module * pipenv introduced for development for easy package management * tests are made pass due to recent changes
