+++ smime/api.py
@@ -1,0 +1,2 @@
+""" """
+from .encrypt import encrypt
